+++ src/exceptions.mjs
@@ -1,1 +1,2 @@
 export { default as PHPJSError } from './exceptions/PHPJSError.mjs';
+export { default as TypePHPJSError } from './exceptions/TypePHPJSError.mjs';