+++ tests/exceptions.global.test.mjs
@@ -3,4 +3,5 @@ import '../src/exceptions.global.js';
 
 test('exceptions.global', () => {
     expect(PHPJSError).not.toBeUndefined();
+    expect(TypePHPJSError).not.toBeUndefined();
 });