+++ tests/exceptions.test.mjs
@@ -1,6 +1,7 @@
 import { expect, test } from '@jest/globals';
-import { PHPJSError } from '../src/exceptions.mjs';
+import { PHPJSError, TypePHPJSError } from '../src/exceptions.mjs';
 
 test('exceptions', () => {
     expect(PHPJSError).not.toBeUndefined();
+    expect(TypePHPJSError).not.toBeUndefined();
 });