+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 -   Добавлен модуль `exceptions` со следующими классами:
 
     -   [PHPJSError](./docs/langref/exceptions/PHPJSError.md) &mdash; Исключение PHPJSError
+    -   [TypePHPJSError](./docs/langref/exceptions/TypePHPJSError.md) &mdash; Исключение
+        TypePHPJSError
 
 -   В модуль `contracts` добавлены следующие интерфейсы: